Run the code on CPU (#4)
* added device instead of hardcoded cuda

* Fix cyclic error

device was set to feats.device and then feats were being move to device. Change to self.device which is the class' device.

* added self.device to feats where missing

* Unnecessary device

torch.backends works even if cuda is not there. Also adapting the training to work on CPU will require more than setting device, it will require changing the lightning flags and adapting the configs to pass those flags appropriately.

* remove line

* Redundant call since logreg and feats are already on device

* all models gets moved to device in line 65, so no need to do it twice

---------

Co-authored-by: Mohamed El Banani <mbanani@umich.edu>